% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_calendar.R
\name{gc_calendar}
\alias{gc_calendar}
\alias{gc_id}
\alias{gc_summary}
\title{Retrieve a Google Calendar}
\usage{
gc_id(x, verbose = TRUE)

gc_summary(x, fixed = FALSE, ..., verbose = TRUE)
}
\arguments{
\item{x}{Calendar-identifying information. In the case of
\code{gc_id}, a character string representing a calendar ID to
which the user has read access. In the case of \code{gc_summary}, a
character string containing a regular expression to be matched to
the calendar title field, \code{summary}.}

\item{verbose}{Logical indicating whether to print informative
messages.}

\item{fixed}{Logical indicating whether \code{x} should be matched
literally.}

\item{\dots}{Optional arguments to be passed to \code{\link{grep}}.}
}
\value{
A \code{googlecalendar} object (a custom class wrapping a
  named list representing metadata for a single calendar).
}
\description{
Retrieves metadata for a Google Calendar in the form of a
\code{googlecalendar} object. This function filters information made
available through the Google Calendar API \code{CalendarList}
resource. Use requires authorization.
}
\details{
For more information on the structure of a \code{CalendarList}
resource (and therefore the expected contents of a
\code{googlecalendar} object), see the Google Calendar API
\href{https://developers.google.com/google-apps/calendar/v3/reference/calendarList}{CalendarList Overview}.
}
\examples{
\dontrun{
gc_id("kpoji8n345cid1eu6o0mlcbs30@group.calendar.google.com")

gc_summary("ments$")
gc_summary("commitments", ignore.case = TRUE)
gc_summary("Commitments", fixed = TRUE)
}

}

