% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_auth.R
\name{gc_deauth}
\alias{gc_deauth}
\title{Deauthorize requests to Google Calendar API}
\usage{
gc_deauth(clear_cache = TRUE, verbose = TRUE)
}
\arguments{
\item{clear_cache}{Logical indicating whether to remove the
\code{.httr-oauth} file in the working directory.}

\item{verbose}{Logical indicating whether to print informative
messages.}
}
\description{
Deauthorizes this package to interact with your Google Calendars.
}
\examples{
\dontrun{
gc_deauth()
}
}

