% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_delete.R
\name{gc_delete}
\alias{gc_delete}
\title{Delete a calendar}
\usage{
gc_delete(x, verbose = TRUE)
}
\arguments{
\item{x}{\code{googlecalendar} object representing the calendar to
delete.}

\item{verbose}{Logical indicating whether to print informative
messages.}
}
\value{
Logical indicating whether the calendar was successfully
  deleted.
}
\description{
Deletes a calendar. This method operates on the Google Calendar API
Calendars resource.
}
\examples{
\dontrun{
gc_summary("Commitments") \%>\%
  gc_delete()
}

}
