% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_edit.R
\name{gc_edit}
\alias{gc_edit}
\title{Edit a calendar}
\usage{
gc_edit(x, ..., verbose = TRUE)
}
\arguments{
\item{x}{\code{googlecalendar} object representing the calendar to
edit.}

\item{verbose}{Logical indicating whether to print informative
messages.}

\item{\dots}{Optional named properties and their new values to be
passed as part of the HTTP request body to the API. May include the
following properties: \code{description}, \code{location},
\code{summary}, \code{timeZone}, \code{backgroundColor},
\code{colorId}, \code{defaultReminders}, \code{foregroundColor},
\code{hidden}, \code{notificationSettings}, \code{selected}, and
\code{summaryOverride}.}
}
\description{
Enables arbitrary edits to the metadata of a calendar. This function
returns an updated \code{googlecalendar} object made available
through the Google Calendar API \code{CalendarList} resource. Use
requires authorization.
}
\details{
This method combines the functionalities of the \code{PUT} methods
for the both the \code{Calendars} and \code{CalendarList} resources.
For more information on properties available for editing, see the
\href{https://developers.google.com/google-apps/calendar/v3/reference/calendars/update}{Calendars: Update}
and
\href{https://developers.google.com/google-apps/calendar/v3/reference/calendarList/update}{CalendarList: Update}
methods overviews in the Google Calendar API reference documentation.
}
\examples{
\dontrun{
cal <- gc_new("Work Schedule")

gc_edit(cal, summary = "Work", location = "New York")
}

}

