% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_event_edit.R
\name{gc_event_edit}
\alias{gc_event_edit}
\title{Edit an event}
\usage{
gc_event_edit(x, ..., verbose = TRUE)
}
\arguments{
\item{x}{\code{event} object representing the event to edit.}

\item{verbose}{Logical indicating whether to print informative
messages.}

\item{\dots}{Optional named properties and their new values to be
passed as part of the HTTP request body to the API. For more
information on available properties, see the
\href{Events: Patch}{https://developers.google.com/google-apps/calendar/v3/reference/events/path}
method overview in the Google Calendar API reference documentation.}
}
\value{
The patched Events resource as an \code{event} object.
}
\description{
Enables arbitrary edits to the metadata of an event. This function
returns an updated \code{event} object made available through the
Google Calendar API \code{Events} resource. Use requires
authorization.
}

