% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_event_import.R
\name{gc_event_import}
\alias{gc_event_import}
\title{Create new events in bulk}
\usage{
gc_event_import(x, events, sendNotifications = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{\code{googlecalendar} object representing the calendar in
which to create the event.}

\item{events}{\code{data.frame} where each row represents an event to
be created.}

\item{sendNotifications}{Logical indicating whether to send
notifications about the event.}

\item{verbose}{Logical indicating whether to print informative
messages.}
}
\value{
Character vector of the newly created event IDs.
}
\description{
Imports a table of new events, adding them to a Google Calendar. If
successful, this function returns a vector of event IDs. Use requires
authorization.
}

