% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_event_ls.R
\name{gc_event_ls}
\alias{gc_event_ls}
\title{List events on a Google Calendar}
\usage{
gc_event_ls(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{A \code{googlecalendar} object representing the calendar for
which to list events.}

\item{verbose}{Logical indicating whether to print informative
messages.}

\item{\dots}{Additional parameters to be passed as part of the HTTP
request to the API.}
}
\value{
An \code{event_ls} object (a custom class wrapping
  \code{\link[dplyr]{tbl_df}}) with one row for each event returned
  by the service.
}
\description{
Lists a user-defined subset of events scheduled on a Google Calendar.
This function returns event information available via the Google
Calendar API \code{Events} resource. Use requires authorization.
}
\details{
Only atomic properties are returned in this listing. To fetch all
properties for an event, use \code{gc_event}. For more information on
fields returned by this function, see the
\href{https://developers.google.com/google-apps/calendar/v3/reference/events}{Events Resource Overview}
in the Google Calendar API reference documentation.
}

