% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_event_new.R
\name{gc_event_new}
\alias{gc_event_new}
\title{Create a new event}
\usage{
gc_event_new(x, start, end, ..., sendNotifications = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{\code{googlecalendar} object representing the calendar in
which to create the event.}

\item{start, end}{Named lists representing the start and end time
properties to be specified on event creation. May include the
following properties: \code{date}, \code{dateTime}, and
\code{timeZone}.}

\item{sendNotifications}{Logical indicating whether to send
notifications about the creation of the new event.}

\item{verbose}{Logical indicating whether to print informative
messages.}

\item{\dots}{Additional parameters to be passed as part of the HTTP request
body to the API. May include the query parameters listed in the
resource documentation. Key properties include: \code{summary},
\code{attendees}, \code{description}, and \code{location}.}
}
\value{
An invisible \code{event} object (a custom class wrapping a
 list) representing metadata for the newly created event.
}
\description{
Creates an event in a Google Calendar. If successful, this function
returns an \code{event} object made available through the Google
Calendar API \code{Events} resource. Use requires authorization.
}
\details{
See \code{\link{events}} for a complete description of the
information made available in an \code{event} object.

For more information on optional event creation properties or fields
returned by this function, see the
\href{https://developers.google.com/google-apps/calendar/v3/reference/events/insert}{Events: Insert}
method overview in the Google Calendar API reference documentation.
}
\examples{
\dontrun{
cal <- gc_summary("Commitments")

gc_event_new(
  cal,
  start = list(dateTime = "2016-06-01T09:00:00Z"),
  end = list(dateTime = "2016-06-01T10:00:00Z"),
  summary = "Lunch with Alyssa"
)
}

}

