% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_ls.R
\name{gc_ls}
\alias{gc_ls}
\title{List calendars a user can view and modify}
\usage{
gc_ls(pattern = NULL, ..., verbose = TRUE)
}
\arguments{
\item{pattern}{Character string containing a regular expression (or
exact character string for \code{fixed = TRUE}) to be matched to
the calendar title field, \code{summary}. If non-\code{NULL}, only
matching calendars will be returned.}

\item{...}{Optional arguments to be passed to \code{\link{grep}}.}

\item{verbose}{Logical indicating whether to print informative
messages.}
}
\value{
A \code{googlecalendar_ls} object (a custom class wrapping a
  \code{\link[dplyr]{tbl_df}}) where each row represents a calendar.
}
\description{
Lists the collection of calendars in the user's calendar list. This
function returns the information available in the Google Calendar API
\code{CalendarList} resource. Use requires authorization.
}
\details{
This listing gives a partial view of the calendars available to be
viewed or modified by the user. Calendars represented in this listing
are the subset of available calendars where \emph{Show in List} has
been selected in
\href{https://calendar.google.com/calendar/render#settings-calendars_9}{Calendar
Settings}.
}
\examples{
\dontrun{
gc_ls()

gc_ls("ts$")
gc_ls("commitments", ignore.case = TRUE)
gc_ls("Commitments", fixed = TRUE)
}

}

