% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_new.R
\name{gc_new}
\alias{gc_new}
\title{Create a new calendar}
\usage{
gc_new(summary, ..., verbose = TRUE)
}
\arguments{
\item{summary}{Character string representing the title of the
calendar.}

\item{verbose}{Logical indicating whether to print informative
messages.}

\item{\dots}{Additional parameters to be passed as part of the HTTP
request body to the API. May include the following properties:
\code{description}, \code{etag}, \code{id}, \code{kind},
\code{location}, and \code{timeZone}.}
}
\value{
A single-row \code{googlecalendar} object (a custom class
  wrapping \code{\link[dplyr]{tbl_df}}) representing metadata for the
  newly created calendar.
}
\description{
Creates a new secondary calendar in the user's list of Google
Calendars. If successful, this function returns a
\code{googlecalendar} object made available through the Google
Calendar API \code{Calendars} resource. Use requires authorization.
}
\details{
See \code{\link{gc_calendar}} for a complete description of the
information made available in a \code{googlecalendar} object.

For more information on optional calendar creation properties or
fields returned by this function, see the
\href{https://developers.google.com/google-apps/calendar/v3/reference/calendars/insert}{Calendars: Insert}
method overview in the Google Calendar API reference documentation.
}
\examples{
\dontrun{
gc_new("R User Group")

gc_new("R User Group",
       description = "Local user group meetings",
       timeZone = "America/Chicago")
}

}

