% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_auth.R
\name{gc_token}
\alias{gc_token}
\title{Generate a Google token}
\usage{
gc_token(file = "googlecalendar_token.rds")
}
\arguments{
\item{file}{character(1) The file path to the 'googlecalendar_token.rds' file}
}
\value{
A \code{httr} request object.
}
\description{
Produces a token for interaction with the Google Calendar API by
either reauthorizing the application or loading a cached token.
}
\keyword{internal}
