% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlecalendar.R
\docType{package}
\name{googlecalendar}
\alias{googlecalendar}
\alias{googlecalendar-package}
\title{googlecalendar}
\description{
Easily access and modify calendars, events, and UI settings via the
Google Calendar API. For more information on installation, usage, and
project development status, see the official
\href{https://github.com/benjcunningham/googlecalendar}{README} on
GitHub.
}

