% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{PATCH_resource}
\alias{PATCH_resource}
\title{Generalized PATCH method}
\usage{
PATCH_resource(path, body)
}
\arguments{
\item{path}{Path of the HTTP request.}

\item{body}{Request body to be encoded as JSON, as a named list.}
}
\value{
If request was successful, an invisible copy of the response.
  In this package, the PATCH return value is rarely used; instead,
  a GET method is usually used to refetch the appropriate object.
}
\description{
Generalized PATCH method
}
\keyword{internal}
