% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{POST_resource}
\alias{POST_resource}
\title{Generalized POST method}
\usage{
POST_resource(path, body, ...)
}
\arguments{
\item{path}{Path of the HTTP request.}

\item{body}{Request body to be encoded as JSON, as a named list.}

\item{\dots}{Additional parameters to be passed to \code{build_url}
as additional named components of the URL query.}
}
\value{
If request was successful, the API response. Otherwise, an
  error or warning, as generated by \code{httr::stop_for_status}.
}
\description{
Generalized POST method
}
\keyword{internal}
