% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_auth.R
\name{token_available}
\alias{token_available}
\title{Check if a Google token exists}
\usage{
token_available()
}
\value{
Logical indicating whether a token exists in the internal
credential environment.
}
\description{
Determines whether a token exists in the internal credential
environment.
}
\keyword{internal}
