% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{DELETE_resource}
\alias{DELETE_resource}
\title{Generalized DELETE method}
\usage{
DELETE_resource(path, ...)
}
\arguments{
\item{path}{Path of the HTTP request.}

\item{\dots}{Additional parameters to be passed to \code{build_url}
as additional named components in the URL query.}
}
\value{
Status code of the API response.
}
\description{
Generalized DELETE method
}
\keyword{internal}
