% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_auth.R
\name{gc_auth}
\alias{gc_auth}
\title{Authorize requests to the Google Calendar API}
\usage{
gc_auth(
  new_user = FALSE,
  key = getOption("googlecalendar.client_key"),
  secret = getOption("googlecalendar.client_secret"),
  cache = getOption("googlecalendar.oauth_cache"),
  token = NULL,
  verbose = FALSE
)
}
\arguments{
\item{new_user}{Logical indicating whether to reauthorize using a new
account and remove the current cached credentials. Currently, only
one account can be authorized or cached in a directory at a time.}

\item{key, secret}{Client key and secret for the application. See
below for more information on setting these in an \code{.Rprofile}
file.}

\item{cache}{Logical indicating whether to cache credentials in a
\code{.httr-oauth} file in the working directory.}

\item{token}{Token object or the name of a token object \code{.rds}
file to be set in the credentialing environment. This is most
useful for authentication in a non-interactive setting.}

\item{verbose}{Logical indicating whether to print informative
messages.}
}
\value{
An OAuth token object.
}
\description{
Authorizes this package to interact with your Google Calendars. You
will be automatically directed to a web browser, asked to sign in to
your Google account, and prompted to grant permission for the package
to interact with the Google Calendar service.
}
\details{
Typically, users will not need to call this function explicitly, as
the authorization process will be triggered by the first action that
requires a user login. However, before using \code{googlecalendar},
the client credentials \code{key} and \code{secret} should be
predefined in an \code{.Rprofile} file using:

\preformatted{
options(
  googlecalendar.client_key = "<KEY>",
  googlecalendar.client_secret = "<SECRET>",
  googlecalendar.oauth_cache = TRUE
)
}

Note that the default value of \code{cache} may also be specified in
the \code{.Rprofile} file using \code{googlecalendar.oauth_cache}.
The recommended value is \code{TRUE}.

For more information on obtaining these credentials, see the project
\href{https://github.com/benjcunningham/googlecalendar}{README} on
GitHub or visit the
\href{https://console.developers.google.com/}{Google API Console} to
get started registering an application.
}
\examples{
\dontrun{
gs_auth()
gs_auth(token = readRDS("saved_token.rds"))
gs_auth(token = "saved_token.rds")
}
}
