% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_calendar.R
\name{gc_calendar}
\alias{gc_calendar}
\alias{gc_id}
\alias{gc_summary}
\title{Retrieve a calendar}
\usage{
gc_id(x, verbose = TRUE)

gc_summary(x, fixed = FALSE, ..., verbose = TRUE)
}
\arguments{
\item{x}{Calendar-identifying information. For \code{gc_id}, a
calendar ID as a character string. For \code{gc_summary}, a regular
expression to be matched to the calendar title field,
\code{summary}.}

\item{verbose}{Logical indicating whether to print informative
messages.}

\item{fixed}{Logical indicating whether \code{x} should be matched
literally.}

\item{\dots}{Optional arguments to be passed to \code{\link[base]{grep}}.}
}
\value{
Calendar metadata as a \code{googlecalendar} object (a custom
  class wrapping a named list).
}
\description{
Retrieves metadata for a Google Calendar in the form of a
\code{googlecalendar} object. This method filters information made
available through the Google Calendar API CalendarList
resource.
}
\details{
While the Google Calendar API prefers to return sparse structures,
\code{googlecalendar} objects attempt to faithfully represent all
documented fields in the CalendarList resource. In practice, this
means that values unreported by the service are reported as \code{NA}
or, if applicable, their documented default values.

For more information on the structure of a \code{googlecalendar}
object, see the Google Calendar API
\href{https://developers.google.com/google-apps/calendar/v3/reference/calendarList}{CalendarList Resource Overview}.
}
\examples{
\dontrun{
gc_id("kpoji8n345cid1eu6o0mlcbs30@group.calendar.google.com")

gc_summary("ments$")
gc_summary("commitments", ignore.case = TRUE)
gc_summary("Commitments", fixed = TRUE)
}

}
