% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_auth.R
\name{gc_deauth}
\alias{gc_deauth}
\title{Deauthorize requests to the Google Calendar API}
\usage{
gc_deauth(clear_cache = TRUE, verbose = TRUE)
}
\arguments{
\item{clear_cache}{Logical indicating whether to remove any
\code{.httr-oauth} file that exists in the working directory.}

\item{verbose}{Logical indicating whether to print informative
messages.}
}
\value{
Logical indicating whether a token was removed from the
  internal credentialing environment.
}
\description{
Deauthorizes this package from interacting with your Google
Calendars.
}
\examples{
\dontrun{
gc_deauth()
}
}
