% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_edit.R
\name{gc_edit}
\alias{gc_edit}
\title{Edit a calendar}
\usage{
gc_edit(x, ..., verbose = TRUE)
}
\arguments{
\item{x}{\code{googlecalendar} object representing the calendar to
edit.}

\item{\dots}{Optional properties and their new values to be passed as
part of the HTTP \code{PATCH} request body.}

\item{verbose}{Logical indicating whether to print informative
messages.}
}
\value{
The edited resource as a \code{googlecalendar} object.
}
\description{
Makes arbitrary edits to the metadata of a calendar. This method
operates on the Google Calendar API CalendarList and Calendars
resources. It filters information made available through the
CalendarList resource.
}
\details{
This method combines the functionalities of the \code{PATCH} methods
for both the Calendars and CalendarList resources. For more
information on properties available for editing, see
\href{https://developers.google.com/google-apps/calendar/v3/reference/calendarList/update}{CalendarList: Update}
and
\href{https://developers.google.com/google-apps/calendar/v3/reference/calendars/update}{Calendars: Update}
in the Google Calendar API reference documentation.
}
\examples{
\dontrun{
gc_new(summary = "Clint Meetings") \%>\%
  gc_edit(summary = "Client Meetings", location = "New York")
}

}
