% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_event_delete.R
\name{gc_event_delete}
\alias{gc_event_delete}
\title{Delete a calendar event}
\usage{
gc_event_delete(x, sendNotifications = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{\code{event} object representing the event to delete.}

\item{sendNotifications}{Logical indicating whether to send
notifications about the event.}

\item{verbose}{Logical indicating whether to print informative
messages.}
}
\value{
Logical indicating whether the event was successfully
  deleted.
}
\description{
Deletes a calendar event. This method operates on the Google Calendar
API Events resource.
}
\examples{
\dontrun{
gc_summary("Commitments") \%>\%
  gc_event_query(q = "Lunch with Mitch") \%>\%
  gc_event_delete()
}

}
