% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_event_edit.R
\name{gc_event_edit}
\alias{gc_event_edit}
\title{Edit a calendar event}
\usage{
gc_event_edit(x, ..., verbose = TRUE)
}
\arguments{
\item{x}{\code{event} object representing the event to edit.}

\item{\dots}{Optional named properties and their new values to be
passed as part of the HTTP request body.}

\item{verbose}{Logical indicating whether to print informative
messages.}
}
\value{
The edited resource as an \code{event} object.
}
\description{
Makes arbitrary edits to the metadata of an event. This method
operates on the Google Calendar API Events resource.
}
\details{
For more information on properties available for editing, see
\href{https://developers.google.com/google-apps/calendar/v3/reference/events/insert}{Events: Insert}.
in the Google Calendar API reference documentation.
}
\examples{
\dontrun{
gc_new("Bar Trivia") \%>\%
  gc_event_new(
    start = list(dateTime = "2016-09-21T20:00:00Z"),
    end = list(dateTime = "2016-09-21T23:30:00Z"),
    summary = "Dons"
  ) \%>\%
  gc_event_edit(summary = "Donnelly's Irish Pub")
}

}
