% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_event_ls.R
\name{gc_event_ls}
\alias{gc_event_ls}
\title{List events in calendar}
\usage{
gc_event_ls(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{A \code{googlecalendar} object representing the calendar for
which to list events.}

\item{\dots}{Additional parameters to be passed as part of the HTTP
request to the API. More information on these named arguments are
available below.}

\item{verbose}{Logical indicating whether to print informative
messages.}
}
\value{
An \code{event_ls} object (a custom class wrapping a
  \code{\link[dplyr]{tbl_df}}) with one row for each event returned
  by the service.
}
\description{
Lists events scheduled on a Google Calendar. This method filters
information made available via the Google Calendar API Events
resource.
}
\details{
An arbitrary number of optional query parameters may be supplied via
\dots. A list of optional query parameters can be found in
\href{https://developers.google.com/google-apps/calendar/v3/reference/events/list#parameters}{Events: List}.

Notable parameters include:

\itemize{
  \item{\code{q} -- Free text search terms to find in any field.}
  \item{\code{singleEvents} -- Logical indicating whether to expand
    recurring events into instances and only return single one-off
    events.}
  \item{\code{timeMax} -- Exclusive upper bound of the start time
    filter (formatted as an RFC3339 timestamp).}
  \item{\code{timeMin} -- Inclusive lower bound of the end time
    filter (formatted as an RFC3339 timestamp).}
}

Note that \code{event_ls} objects only represent resource properties
that are naturally atomic. To fetch all properties for an event, use
\code{gc_event}. For more information on the structure of an
\code{event_ls} object, see the Google Calendar API
\href{https://developers.google.com/google-apps/calendar/v3/reference/events}{Events Resource Overview}.
}
\examples{
\dontrun{
gc_summary("Commitments") \%>\%
  gc_event_ls()
}

}
