% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_event_new.R
\name{gc_event_new}
\alias{gc_event_new}
\title{Create a new calendar event}
\usage{
gc_event_new(x, start, end, ..., sendNotifications = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{\code{googlecalendar} object representing the calendar in
which to create the event.}

\item{start, end}{Named list representing the start or end time. May
include the following properties: \code{date}, \code{dateTime}, and
\code{timeZone}.}

\item{\dots}{Additional parameters to be passed as part of the HTTP
request body. More information on these optional values is
available below.}

\item{sendNotifications}{Logical indicating whether to send
notifications about the event.}

\item{verbose}{Logical indicating whether to print informative
messages.}
}
\value{
Event metadata as an invisible \code{event} object (a custom
  class wrapping a named list).
}
\description{
Creates an event in a Google Calendar. This method operates on the
Google Calendar API Events resource.
}
\details{
For more information on the structure of an \code{event} object, see
the Google Calendar API
\href{https://developers.google.com/google-apps/calendar/v3/reference/events}{Events Resource Overview}.
}
\examples{
\dontrun{
cal <- gc_summary("Commitments")

gc_event_new(
  cal,
  start = list(dateTime = "2016-06-01T09:00:00Z"),
  end = list(dateTime = "2016-06-01T10:00:00Z"),
  summary = "Lunch with Alyssa"
)
}

}
