% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_new.R
\name{gc_new}
\alias{gc_new}
\title{Create a new calendar}
\usage{
gc_new(summary, ..., verbose = TRUE)
}
\arguments{
\item{summary}{Character string representing the title of the
calendar.}

\item{\dots}{Additional parameters to be passed as part of the HTTP
request body. More information on these optional values is
available below.}

\item{verbose}{Logical indicating whether to print informative
messages.}
}
\value{
Calendar metadata as an invisible \code{googlecalendar}
  object (a custom class wrapping a named list).
}
\description{
Creates a new secondary calendar. This method operates on the Google
Calendar API Calendars resource.
}
\details{
For more information on the structure of a calendar object, see the
Google Calendar API
\href{https://developers.google.com/google-apps/calendar/v3/reference/calendarList}{CalendarList Resource Overview}.
}
\examples{
\dontrun{
gc_new(
  "R User Group",
  description = "Local user group meetings",
  timeZone = "America/Chicago"
)
}

}
