% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{GET_resource}
\alias{GET_resource}
\title{Generalized GET method}
\usage{
GET_resource(path, fields = NULL)
}
\arguments{
\item{path}{Path of the HTTP request.}

\item{fields}{Fields to be included in the response, as a character
vector.}
}
\value{
If request was successful, the API response. Otherwise, an
  error or warning, as generated by \code{httr::stop_for_status}.
}
\description{
Generalized GET method
}
\keyword{internal}
